update dir struct
