+++ examples/dependency_parser/ddparser/criterion.py
@@ -1,0 +1,43 @@
+# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import numpy as np
+import paddle
+import paddle.nn as nn
+import paddle.nn.functional as F
+
+from utils import index_sample
+
+
+class ParserCriterion(nn.Layer):
+    def __init__(self, *args, **kwargs):
+        super(ParserCriterion, self).__init__(*args, **kwargs)
+
+    def forward(self, s_arc, s_rel, arcs, rels, mask):
+
+        arcs = paddle.masked_select(arcs, mask)
+        rels = paddle.masked_select(rels, mask)
+
+        select = paddle.nonzero(mask)
+        s_arc = paddle.gather_nd(s_arc, select)
+        s_rel = paddle.gather_nd(s_rel, select)
+
+        s_rel = index_sample(s_rel, paddle.unsqueeze(arcs, axis=1))
+
+        arc_cost = F.cross_entropy(s_arc, arcs)
+        rel_cost = F.cross_entropy(s_rel, rels)
+
+        avg_cost = paddle.mean(arc_cost + rel_cost)
+        return avg_cost
+
